+++ src/cache/authLink.js
@@ -1,0 +1,15 @@
+import { setContext } from 'apollo-link-context';
+
+const authLink = setContext((_, { headers }) => {
+  // get the authentication token from local storage if it exists
+  const token = localStorage.getItem('jwt');
+  // return the headers to the context so httpLink can read them
+  return {
+    headers: {
+      ...headers,
+      Authorization: token ? `JWT ${token}` : '',
+    },
+  };
+});
+
+export default authLink;
