+++ src/cache/cache.js
@@ -1,0 +1,3 @@
+import { InMemoryCache } from 'apollo-cache-inmemory';
+
+export default new InMemoryCache();
