+++ src/cache/httpLink.js
@@ -1,0 +1,7 @@
+import { HttpLink } from 'apollo-link-http';
+
+const httpLink = new HttpLink({
+  uri: 'http://potato-store.herokuapp.com/graphql',
+});
+
+export default httpLink;
