+++ src/components/FilterMenu/MinMaxSelect.css
